Modify the structure of the class DataEntry.
Enhanced Semantic Expression of the DataEntry.In order to better characterize KV data structure, key and headers fields is added in DataEntry. At the same time, the function of value checking and string parsing is also added. Better yet, it provide basic and custom types for user selection and extension.

 The old structure as the STRUCT type of the new structure, you can use it as new way and more easier.

 We change SCHEMA to META because we think META is better than SCHEMA in approaching the meaning expressed by this object.

 We added the headers attribute to correspond to userProperty in Massage of RocketMQ.

 We provide a variety of static construction methods and static variables for all type of metas, to facilitate users to quickly select the required meta types. We also provides a fast conversion method for object data, so that users can get the required data safely and quickly.